requirements.txt
# Web Framework
Flask==3.1.1
Flask-CORS==6.0.1
Werkzeug==3.1.3
gradio==5.47.0

# Database
psycopg2-binary==2.9.10

# Vector Database
chromadb==1.0.15

# Machine Learning
sentence-transformers==5.0.0
transformers==4.53.2
Pillow==11.3.0

# AI/LLM
ollama==0.5.3
llama-index==0.13.5
llama-index-llms-ollama==0.7.1

# Hugging Face Hub for model loading
huggingface-hub==0.35.1

# Utilities
numpy==2.3.1
pandas==2.2.3
python-dotenv==1.1.1
gunicorn==21.2.0
nest-asyncio==1.6.0
tqdm==4.67.1
gdown==4.7.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0